refactor: moved activeAntelopeChain to assetDetails
