+++ src/containers/LoginPassphraseContainer.tsx
@@ -14,7 +14,8 @@ import useErrorStore from '../store/errorStore';
 import { DEFAULT_DEV_PASSPHRASE_LIST } from '../store/passphraseStore';
 import PassphraseInput from '../components/PassphraseInput';
 import { createNetworkErrorState, isNetworkError } from '../utils/errors';
-import { activeAntelopeChainEntry, AntelopeAccount, AntelopePrivateKey } from '../utils/chain/antelope';
+import { AntelopeAccount, AntelopePrivateKey } from '../utils/chain/antelope';
+import { activeAntelopeChainEntry } from '../utils/assetDetails';
 import { Asset } from '../utils/chain/types';
 import { assetStorage, keyStorage } from '../utils/StorageManager/setup';
 import Debug from 'debug';
